RaspberryPI/requirements.txt
azure-iot-device==1.18.3
matplotlib==3.8.2
mpu6050==0.0.3
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
Pillow==10.3.0
pynmea2==1.19.0
pyserial==3.5
torch==2.2.0
torchvision==0.17.0
smbus2==0.4.3